pyproject.toml
[project]
name = "pave"
version = "0.1.0"
description = "PAVE RCA Agent - Root Cause Analysis agent using LangGraph"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiolimiter>=1.2.1",
    "colorlog>=6.9.0",
    "datasets>=4.8.4",
    "duckdb>=1.4.3",
    "huggingface-hub>=1.11.0",
    "langchain>=1.2.6",
    "langchain-anthropic>=1.3.1",
    "langchain-google-genai>=4.2.0",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.6",
    "polars>=1.40.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "pytz>=2026.1.post1",
    "pyyaml>=6.0.2",
    "rcabench-platform[llm-eval]>=0.4.39",
    "tenacity>=8.2.0",
]

[project.entry-points."llm_eval.agents"]
pave = "pave.agents.eval_agent:PAVEAgent"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pave"]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
addopts = ["-v", "-s"]

[tool.uv]
index-url = "https://pypi.org/simple"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
isort = { combine-as-imports = true, known-first-party = ["pave"] }

[tool.ruff.lint.pydocstyle]
convention = "google"

[dependency-groups]
dev = ["pytest>=9.0.2", "pytest-asyncio>=1.3.0", "ruff>=0.15.7"]
